requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
# Removemos a linha do httpx para evitar o conflito
supabase>=2.3.0
redis==5.0.1
celery==5.3.6
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.7
langchain-community==0.0.24
qdrant-client==1.7.3
openai==1.12.0
tiktoken==0.6.0